Do away with classes, use plain functions
NDISC-68
