+++ time-machine/src/main/kotlin/de/bund/digitalservice/ris/norms/timemachine/TimeMachine.kt
@@ -6,13 +6,10 @@ import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.types.file
 import de.bund.digitalservice.ris.norms.timemachine.core.applyModification
-import de.bund.digitalservice.ris.norms.timemachine.helper.FileToDocumentConverter
-import de.bund.digitalservice.ris.norms.timemachine.helper.XmlFileWriter
-import java.io.StringWriter
-import javax.xml.transform.TransformerFactory
-import javax.xml.transform.dom.DOMSource
-import javax.xml.transform.stream.StreamResult
+import documentToString
+import fileToDocument
 import org.w3c.dom.Document
+import writeDocumentToFile
 
 class TimeMachine : CliktCommand() {
   private val amendingLawFile by argument().file(mustExist = true, canBeDir = false)
@@ -20,8 +17,8 @@ class TimeMachine : CliktCommand() {
   private val doPrintToStandardOutput by option("--stdout").flag()
 
   override fun run() {
-    val amendingLawDoc: Document = FileToDocumentConverter().convert(amendingLawFile)
-    val targetLawDoc = FileToDocumentConverter().convert(targetLawFile)
+    val amendingLawDoc: Document = fileToDocument(amendingLawFile)
+    val targetLawDoc = fileToDocument(targetLawFile)
 
     val appliedLaw = applyModification(amendingLawDoc, targetLawDoc)
 
@@ -30,20 +27,8 @@ class TimeMachine : CliktCommand() {
       echo(documentString, true, false)
     }
 
-    XmlFileWriter()
-        .writeDocumentToFile(appliedLaw, targetLawFile.nameWithoutExtension + "_amended.xml")
+    writeDocumentToFile(appliedLaw, targetLawFile.nameWithoutExtension + "_amended.xml")
   }
-}
-
-// TODO: move to a different place
-fun documentToString(document: Document): String {
-  val domSource = DOMSource(document)
-  val writer = StringWriter()
-  val result = StreamResult(writer)
-  val transformerFactory = TransformerFactory.newInstance()
-  val transformer = transformerFactory.newTransformer()
-  transformer.transform(domSource, result)
-  return writer.toString()
 }
 
 fun main(args: Array<String>) = TimeMachine().main(args)
